requested changes made
