+++ src/ChainSafe.Gaming.HyperPlay/Dto/Chain.cs
@@ -1,0 +1,10 @@
+using Newtonsoft.Json;
+
+namespace ChainSafe.Gaming.HyperPlay.Dto
+{
+    public struct Chain
+    {
+        [JsonProperty("chainId")]
+        public string ChainId { get; set; }
+    }
+}
