+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.5.0/Web3.Unity Samples/Scripts/Scenes/HyperPlayLoginProvider.cs
@@ -26,7 +26,7 @@ public class HyperPlayLoginProvider : LoginProvider, IWeb3BuilderServiceAdapter
     {
         return web3Builder.Configure(services =>
         {
-            services.UseHyperPlay();
+            services.UseHyperPlay().UseHyperPlaySigner().UseHyperPlayTransactionExecutor();
         });
     }
